+++ contracts/core/exchange-wrappers/ZeroExExchangeWrapper.sol
@@ -85,7 +85,7 @@ contract ZeroExExchangeWrapper {
      */
     function exchange(
         address _taker,
-        uint _orderCount,
+        uint256 _orderCount,
         bytes _orderData
     )
         external
