+++ contracts/core/interfaces/ISetToken.sol
@@ -30,12 +30,12 @@ interface ISetToken {
     /*
      * Get natural unit of Set
      *
-     * @return  uint       Natural unit of Set
+     * @return  uint256       Natural unit of Set
      */
     function naturalUnit()
         external
         view
-        returns (uint);
+        returns (uint256);
 
     /*
      * Get addresses of all components in the Set
@@ -55,7 +55,7 @@ interface ISetToken {
     function getUnits()
         external
         view
-        returns(uint[]);
+        returns(uint256[]);
 
     /*
      * Mint set token for given address.
@@ -66,7 +66,7 @@ interface ISetToken {
      */
     function mint(
         address _issuer,
-        uint _quantity
+        uint256 _quantity
     )
         external;
 
@@ -79,7 +79,7 @@ interface ISetToken {
      */
     function burn(
         address _from,
-        uint _quantity
+        uint256 _quantity
     )
         external;
 }
